idée profit sans observer pattern

+++ booking-api/src/main/java/ca/ulaval/glo4002/booking/domain/Price.java
@@ -30,8 +30,8 @@ public class Price {
         return price.getAmount();
     }
 
-    public double getRoundedAmount(int numberOfDecimals) {
-        return price.rounded(numberOfDecimals, ROUNDING_MODE).getAmount().doubleValue();
+    public float getRoundedAmount(int numberOfDecimals) {
+        return price.rounded(numberOfDecimals, ROUNDING_MODE).getAmount().floatValue();
     }
 
     @Override
